Refactoring functions with program string arguments; drop suffix _P.
1. Rename IOStream::print_P, printf_P, puts_P; drop suffix.
2. Ripple effect on drivers and example sketches.

+++ examples/Ethernet/CosaSNMP/CosaSNMP.ino
@@ -117,7 +117,7 @@ ARDUINO_MIB::is_request(SNMP::PDU& pdu)
   if(pdu.oid.length > (mib_baselen + 2)) return (false);
 
   // Match with Arduino MIB OID root
-  uint8_t index = pdu.oid.match(get_oid());
+  int index = pdu.oid.match(get_oid());
   if (index < -1) return (false); // MIB is greater than what we deal with
   uint8_t sys = (pdu.oid.length > mib_baselen) ? pdu.oid.name[mib_baselen] : 0;
   // Get next value; adjust index referens
